examples/flask_auth_pynamodb/requirements.txt
Flask==0.11.1
Flask-Cors==3.0.2
Flask-GraphQL==1.3.0
Flask-JWT==0.3.2
graphene==1.0.2
graphene-pynamodb==0.3
